Move reader/writer logic to F#

+++ UnoCash.Dto/Expense.fs
@@ -4,7 +4,7 @@ open System
 
 type Expense =
     {
-        Amount      : decimal 
+        Amount      : float 
         Payee       : string  
         Description : string  
         Date        : DateTime
